add playback/deviceList getter
